refactor: 테스트에서만 사용하는 클래스 이름, 위치 변경 및 spring data 사용을 위한 빈 메서드 구현
